+++ src/srv/srv.h
@@ -6,7 +6,7 @@
 /*   By: kcharla <kcharla@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/23 23:46:25 by kcharla           #+#    #+#             */
-/*   Updated: 2020/11/02 16:34:50 by kcharla          ###   ########.fr       */
+/*   Updated: 2020/11/02 20:26:15 by kcharla          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,12 @@
 # define MSG_OK     1
 # define MSG_ACTIVE 2
 # define MSG_EXIT   3
+# define MSG_SHUT   4
 # define MSG_ERROR -1
+
+# define SRV_HELLO "Server is ready. Type 'exit' to close connection.\n"
+# define SRV_BUSY  "Server is currently used by another session/user. Try later.\n"
+# define SRV_ERR   "Some error caused server to close connection. Press 'Enter' to continue.\n"
 
 /*
 ** srv_process.c
@@ -35,10 +40,10 @@ int					srv_process_line(t_srv *srv, char *str, char **response);
 int					srv_init(t_rt *rt);
 
 /*
-** srv_exit.c
+** srv_shutdown.c
 */
 
-int					srv_exit(t_rt *rt);
+int					srv_shutdown(t_rt *rt);
 
 /*
 ** srv_loop.c
@@ -56,15 +61,19 @@ int					srv_request(t_rt *rt, char *str);
 ** srv_ext.c
 */
 
-int					srv_ext_client_check(t_srv *srv);
 int					srv_ext_client_process(t_rt *rt);
+int					srv_ext_client_check(t_srv *srv);
 int					srv_ext_client_disconnect(t_srv *srv);
+
+/*
+** srv_ext_data.c
+*/
+
 int					srv_ext_client_get_data(t_srv *srv);
 int					srv_ext_client_update_str(t_srv *srv);
-int					srv_ext_client_str_process(t_srv *srv);
-
-
-int					srv_process_line(t_srv *srv, char *str, char **response);
+int					srv_ext_client_parse(t_srv *srv);
+int					srv_ext_client_str_parse(t_srv *srv, unsigned long *line_begin);
+int					srv_ext_client_parse_helper(t_srv *srv, unsigned long line_begin);
 
 /*
 ** srv_parse.c
