+++ src/srv/srv_shutdown.c
@@ -1,23 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   srv_exit.c                                         :+:      :+:    :+:   */
+/*   srv_shutdown.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kcharla <kcharla@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/10/31 19:28:56 by kcharla           #+#    #+#             */
-/*   Updated: 2020/10/31 19:30:56 by kcharla          ###   ########.fr       */
+/*   Created: 2020/11/02 20:25:43 by kcharla           #+#    #+#             */
+/*   Updated: 2020/11/02 20:25:56 by kcharla          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "rt.h"
+
 /*
 ** Must be run from gtk thread only!
 */
 
-int		srv_exit(t_rt *rt)
+int		srv_shutdown(t_rt *rt)
 {
 	if (rt == NULL)
 		return (rt_err("rt is NULL pointer"));
-	return (srv_request(rt, "exit"));
+	return (srv_request(rt, "shutdown"));
 }
