+++ src/srv/srv_types.h
@@ -6,7 +6,7 @@
 /*   By: kcharla <kcharla@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/01 16:47:19 by kcharla           #+#    #+#             */
-/*   Updated: 2020/11/02 15:49:25 by kcharla          ###   ########.fr       */
+/*   Updated: 2020/11/02 17:58:47 by kcharla          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
